@@ -9,6 +9,9 @@ import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
 import { TaskService } from './task.service';
 import { ClickStopPropagationDirective } from './directive/stop-click/click-stop-propagation.directive';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { MatTableModule } from '@angular/material/table';
+import { MatSortModule } from '@angular/material/sort';
 
 @NgModule({
   declarations: [AppComponent, ClickStopPropagationDirective],
@@ -16,7 +19,11 @@ import { ClickStopPropagationDirective } from './directive/stop-click/click-stop
     BrowserModule,
     IonicModule.forRoot(),
     AppRoutingModule,
-    // HttpClientModule,
+    HttpClientModule,
+    BrowserAnimationsModule,
+    MatTableModule,
+    MatSortModule,
+
   ],
   providers: [
     { provide: RouteReuseStrategy, useClass: IonicRouteStrategy },
